+++ shape/label.go
@@ -6,6 +6,7 @@ import (
 	"io"
 
 	"github.com/gregoryv/draw/xy"
+	"github.com/gregoryv/nexus"
 )
 
 func NewLabel(text string) *Label {
@@ -21,7 +22,9 @@ type Label struct {
 	x int
 	y int
 
-	Text  string
+	Text string
+	href string
+
 	Font  Font
 	Pad   Padding
 	class string
@@ -30,6 +33,8 @@ type Label struct {
 func (l *Label) String() string {
 	return fmt.Sprintf("label %s at %v", l.Text, &xy.Point{X: l.x, Y: l.y})
 }
+
+func (l *Label) SetHref(v string) { l.href = v }
 
 func (l *Label) Position() (int, int) {
 	return l.x, l.y
@@ -45,13 +50,20 @@ func (l *Label) Height() int          { return l.Font.LineHeight }
 func (l *Label) Direction() Direction { return DirectionRight }
 func (l *Label) SetClass(c string)    { l.class = c }
 
-func (l *Label) WriteSVG(w io.Writer) error {
+func (l *Label) WriteSVG(out io.Writer) error {
 	x, y := l.Position()
 	y += l.Font.LineHeight
-	_, err := fmt.Fprintf(w,
-		`<text class="%s" font-size="%vpx" x="%v" y="%v">%s</text>`,
+	w, err := nexus.NewPrinter(out)
+
+	if l.href != "" {
+		w.Printf(`<a href="%s">`, l.href)
+	}
+	w.Printf(`<text class="%s" font-size="%vpx" x="%v" y="%v">%s</text>`,
 		l.class, l.Font.Height, x, y, l.Text)
-	return err
+	if l.href != "" {
+		w.Printf(`</a>`)
+	}
+	return *err
 }
 
 func (l *Label) Edge(start xy.Point) xy.Point {